Fix CollisionRegion serialization/deserialization issues and CollisionPick implicit cast warning

+++ interface/src/raypick/CollisionPick.cpp
@@ -205,7 +205,7 @@ void CollisionPick::computeShapeInfo(CollisionRegion& pick, ShapeInfo& shapeInfo
                 for (const FBXMeshPart& meshPart : mesh.parts) {
                     triangleIndicesCount += meshPart.triangleIndices.count();
                 }
-                triangleIndices.reserve(triangleIndicesCount);
+                triangleIndices.reserve((int)triangleIndicesCount);
 
                 for (const FBXMeshPart& meshPart : mesh.parts) {
                     const int* indexItr = meshPart.triangleIndices.cbegin();
